third_party/blink/web_tests/storage/indexeddb/mozilla/create-objectstore-null-name-expected.txt

Test IndexedDB's creating object store with null name

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "create-objectstore-null-name.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore(null);
PASS objectStore.name is 'null'
PASS successfullyParsed is true

TEST COMPLETE
